Move camera controller and add key bindings
