--- Assets/Scripts/Strategy/StaticMap.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 4c58e19b76648594e9152c053e487bdd
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
